requirements.txt
aiogram
pandas
openpyxl
python-dotenv
httpx
openai
